Dmi2sheet Refactoring
The code is much, much cleaner now. By default, this only works for linear-to-grid. I also fixed some bugs with copying to a grid destination.

I need to now break the linear chunking down to its own function for extracting.  Then do the same for chunking from the destination chunk.